Rename the ro_json_parser_tests.c file + add equality tests for dicts
